pyproject.toml
[tool.poetry]
name        = "TFM"
version     = "0.1.0"
package-mode = false
description = "TFM"
authors     = ["TheSOV <sov.tele@gmail.com>"]
readme      = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
pyyaml = "6.0.2"
asyncio = "3.4.3"
litellm = "1.68.0"
python-dotenv = "1.1.0"
weaviate-client = "4.14.1"
langchain = "0.3.25"
sentence-transformers = "4.1.0"
wheel = "0.45.1"
einops = "0.8.1"
deepdiff = "8.5.0"
checkov = "3.2.424"
kubernetes-validate = "1.32.0"
yamllint = "1.37.1"
massedit = "^0.70.0"
python-dxf = "^12.1.1"
kubernetes = "^32.0.1"
mlflow = "^2.22.1"
dulwich = "^0.22.8"
crewai = "^0.126.0"
crewai-tools = "^0.47.0"
tenacity = "^9.1.2"
pytest = "^8.4.0"
crawl4ai = "^0.6.3"
flask = "^3.1.1"
flask-cors = "^6.0.1"
gunicorn = "^23.0.0"

[[tool.poetry.source]]
name     = "pytorch-cu118"
url      = "https://download.pytorch.org/whl/cu118"
priority = "explicit"

[tool.poetry.group.cu118.dependencies]
torch = { version = "2.7.0+cu118", source = "pytorch-cu118" }

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
